requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-autoslug==1.9.9
django-cities-light==3.10.2
idna==3.10
pillow==11.3.0
progressbar2==4.5.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-utils==3.9.1
pytz==2025.2
requests==2.32.5
shortuuid==1.0.13
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.5.0
gunicorn==23.0.0
whitenoise==6.7.0
